requirements.txt
tensorflow
keras
keras_vggface
Pillow
numpy
scipy
opencv-python
